Added difficulty sliders in the game start menu and made them affect enemies and scoring

+++ Assets/Scripts/Enemy/EnemyAttack.cs
@@ -6,7 +6,7 @@ namespace CompleteProject
     public class EnemyAttack : MonoBehaviour
     {
         public float timeBetweenAttacks = 0.5f;
-        public int attackDamage = 10;
+        public float attackDamage = 10;
 
 
         Animator anim;
