+++ Assets/Scripts/Managers/ScoreManager.cs
@@ -6,7 +6,7 @@ namespace CompleteProject
 {
     public class ScoreManager : MonoBehaviour
     {
-        public static int score;
+        public static float score;
         public static int kills;
         public Text scoreText;
